*db:     * move models to package     !* relations     ! table naming + gitignore * isolated pydantic models +*! routers *! exceptions annotations
++ todo's
