+++ pars.py
@@ -2,8 +2,18 @@ from selenium import webdriver
 from selenium.webdriver.chrome.options import Options
 from selenium.webdriver.chrome.service import Service
 from selenium.webdriver.common.by import By
-from database import Product, add_item
 
+from models.db import Product
+
+
+# def add_item(data: Product):
+#     exists_name = SessionLocal.query(Product.name).filter(Product.name == data.name)
+#     if not SessionLocal.query(exists_name.exists()).scalar():
+#         SessionLocal.add(data)
+#         SessionLocal.commit()
+#         print("Товар добавлен")
+#     elif SessionLocal.query(Product.price).filter(Product.price == data.price):
+#         print("Цена не изменилась")
 
 
 def get_price(url):
