+++ routes/__init__.py
@@ -1,0 +1,2 @@
+from .users import users_router
+from .products import products_router
